ollama_agent/pyproject.toml
[project]
name = "ollama-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alex Tan Hong Pin", email = "alextan220990@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.2.0",
    "litellm>=1.67.0",
]

[project.scripts]
ollama-agent = "ollama_agent:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
